fixes per mani's comments
