+++ play/src/components/joingame/howtoplayslides/HowToPlaySlide4Content.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { useTheme } from '@mui/material/styles';
+import { Typography } from '@mui/material';
+import {
+  OverlayContainerStyled,
+  OverlayImageStyled,
+  ScreenshotImageStyled,
+} from '../../../lib/styledcomponents/CarouselElementsStyled';
+import HowToPlay_BlueMonster from '../../../img/HowToPlay_BlueMonster.svg';
+import HowToPlay_Screenshot4 from '../../../img/HowToPlay_Screenshot4.svg';
+
+export default function HowToPlaySlide4Content() {
+  const theme = useTheme();
+
+  return (
+    <>
+      <OverlayContainerStyled>
+        <OverlayImageStyled
+          src={HowToPlay_BlueMonster}
+          alt="monster"
+          sx={{
+            top: '30px',
+            left: '370px',
+            width: '140px',
+            height: 'auto',
+            zIndex: -1,
+          }}
+        />
+      </OverlayContainerStyled>
+      <ScreenshotImageStyled src={HowToPlay_Screenshot4} alt="monster" />
+      <Typography
+        variant="h2"
+        sx={{
+          textAlign: 'center',
+          paddingTop: `${theme.sizing.mediumPadding}px`,
+          width: '350px',
+        }}
+      >
+        The most total points wins!
+      </Typography>
+    </>
+  );
+}
